add standardized tests
