+++ my-nextjs-app/puppeteer-test-results.json
@@ -1,16 +1,1 @@
-[
-  {
-    "testName": "standard idiomatic nextjs",
-    "timeToButtonVisible": 28.991371005773544,
-    "timeButtonTookToRespond": 26128.43629400432,
-    "hadToWaitForNetworkIdleAndClickAgain": true,
-    "secondButtonClickResponseTime": 35.16081500053406
-  },
-  {
-    "testName": "optimized using direct DOM API nextjs",
-    "timeToButtonVisible": 25.06740900874138,
-    "timeButtonTookToRespond": 10.80248099565506,
-    "hadToWaitForNetworkIdleAndClickAgain": false,
-    "secondButtonClickResponseTime": false
-  }
-]
+[]
