+++ my-qwik-app/docker-compose.yml
@@ -34,8 +34,9 @@ services:
     cap_add:
       - SYS_ADMIN
     volumes:
+      - ../standardized-puppeteer-test.js:/home/pptruser/standardized-puppeteer-test.js
       - ./puppeteer-test.js:/home/pptruser/app/puppeteer.js
-      - ./puppeteer-test-results.json:/home/pptruser/app/puppeteer-test-results.json
+      - ./puppeteer-test-results.json:/home/pptruser/puppeteer-test-results.json
     command: node app/puppeteer.js
     restart: "no"
     depends_on:
