+++ my-qwik-app/runTest.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+chmod 777 puppeteer-test-results.json
+
 docker compose up --exit-code-from puppeteer
 EXIT_CODE=$?
