+++ my-qwik-app/src/routes/counter2/index.tsx
@@ -1,0 +1,22 @@
+import { component$ } from "@builder.io/qwik";
+import type { DocumentHead } from "@builder.io/qwik-city";
+
+import Counter from "../../components/starter/counter/counter2";
+
+export default component$(() => {
+  return (
+    <>
+      <Counter />
+    </>
+  );
+});
+
+export const head: DocumentHead = {
+  title: "Welcome to Qwik",
+  meta: [
+    {
+      name: "description",
+      content: "Qwik site description",
+    },
+  ],
+};
