+++ rezact/puppeteer-test.js
@@ -1,59 +1,14 @@
 const puppeteer = require("puppeteer");
 const fs = require("fs");
+const {
+  runStandardTest,
+  saveReport,
+} = require("../standardized-puppeteer-test");
 
-fs.writeFileSync("app/puppeteer-test-results.json", JSON.stringify({}));
+async function run() {
+  await runStandardTest("Rezact", "http://proxy:9173/");
 
-puppeteer.launch({ headless: "new" }).then(async (browser) => {
-  const page = await browser.newPage();
-
-  let pageLoadStart = performance.now();
-  console.log("goto Page");
-  page.goto("http://proxy:9173/");
-
-  console.log("wait for button");
-  let btn = await page.waitForSelector("#test-inc");
-  const buttonVisible = performance.now();
-
-  console.log("wait 1 second like a human might to click the button");
-  await sleep(1000);
-
-  console.log("click button");
-  btn = await page.waitForSelector("#test-inc");
-  const startTimer = performance.now();
-  btn.click();
-
-  console.log("wait for value 1");
-  let waiting = true;
-  let counter = 0;
-  while (waiting) {
-    counter++;
-    // await sleep(1);
-    const results = await page.evaluate(() => {
-      const valElm = document.querySelector("#test-value");
-      if (valElm) return valElm.textContent;
-    });
-    if (results === "1") waiting = false;
-    if (counter > 1000) throw new Error("timeout");
-  }
-
-  const endTimer = performance.now();
-
-  console.log("time diff", endTimer - startTimer);
-  fs.writeFileSync(
-    "app/puppeteer-test-results.json",
-    JSON.stringify(
-      {
-        timeToButtonVisible: buttonVisible - pageLoadStart,
-        timeButtonTookToRespond: endTimer - startTimer,
-      },
-      null,
-      2,
-    ),
-  );
-
-  await browser.close();
-});
-
-async function sleep(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
+  saveReport();
 }
+
+run();
